Langchain_prompts/1_prompt_ui.py
from langchain_huggingface import ChatHuggingFace, HuggingFaceEndpoint
from dotenv import load_dotenv
import streamlit as st
import os

load_dotenv()


load_dotenv()

repo_id1 = "mistralai/Mixtral-8x7B-Instruct-v0.1"

llm = HuggingFaceEndpoint(
    repo_id=repo_id1,
    task="text-generation", 
)

model = ChatHuggingFace(llm=llm)

st.header("Research Tool")

user_input = st.text_input("Enter your prompt")


if st.button("Summarize"):
    result = model.invoke(user_input)
    st.write(result.content)
    